requirements.txt
google-genai==1.28.0
pytest>=8.0.0
requests>=2.32.0